+++ src/utils/mod.rs
@@ -1,7 +1,7 @@
 pub mod direction;
 pub mod point;
 pub mod rotation;
-pub mod weight_indexed_graph;
+pub mod source_sink_graph;
 
 pub fn set_panic_hook() {
     // When the `console_error_panic_hook` feature is enabled, we can call the
